extract application code to another file
