+++ lib/index.js
@@ -5,9 +5,7 @@ const {
   always,
   assoc,
   find,
-  forEach,
   isEmpty,
-  map,
   merge,
   pickAll,
   pipe,
@@ -16,32 +14,9 @@ const {
   toString,
   when,
 } = require('ramda')
+const { createApplication } = require('./application')
 
-module.exports = function createApplication () {
-  const application = {
-    routes: [],
-  }
-
-  const buildMethodRoute = app => method =>
-    (path, handler) => {
-      app.routes.push({
-        path, handler, method
-      })
-    }
-
-  const addMethodRouteToApp = app => methodName =>
-    app[methodName] = buildMethodRoute(app)(methodName)
-
-  pipe(
-    map(toLower),
-    forEach(addMethodRouteToApp(application))
-  )(http.METHODS)
-
-  application.listen = PORT =>
-    createServer(application.routes).listen(PORT)
-
-  return application
-}
+module.exports = createApplication(createServer)
 
 function createServer (routes) {
   const httpHandler = (httpReq, httpRes) => {
